Factor initial_processing.py into process and eap pieces.
