+++ quipucords/scanner/network/inspect_callback.py
@@ -15,7 +15,7 @@ from django.db import transaction
 from ansible.plugins.callback import CallbackBase
 from api.models import (ScanTask, InspectionResult,
                         SystemInspectionResult, RawFact)
-from scanner.network import initial_processing
+from scanner.network.processing import process
 
 # Get an instance of a logger
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
@@ -90,7 +90,7 @@ class InspectResultCallback(CallbackBase):
     @transaction.atomic
     def _finalize_host(self, host):
         facts = self._ansible_facts.get(host, {})
-        results = initial_processing.process(facts)
+        results = process.process(facts)
 
         logger.debug('host scan complete for %s with facts %s',
                      host, results)
@@ -110,7 +110,7 @@ class InspectResultCallback(CallbackBase):
 
         # Generate facts for host
         for result_key, result_value in results.items():
-            if result_value == initial_processing.NO_DATA:
+            if result_value == process.NO_DATA:
                 continue
 
             stored_fact = RawFact(name=result_key,
