+++ quipucords/scanner/network/processing/eap.py
@@ -9,133 +9,12 @@
 
 """Initial processing of raw shell output from Ansible commands."""
 
-import abc
-import json
 import logging
+from scanner.network.processing import process
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
-# ### Conventions ####
-#
-# The processing functions return strings or None because that's what
-# can fit in a RawFact in our database. Normal results are
-# JSON-encoded strings. In case of error, the result is '' and the
-# error is logged.
-
-# #### Infrastructure ####
-
-PROCESSORS = {}
-NO_DATA = ''  # Result value when we have errors.
-
-
-def process(facts):
-    """Do initial processing of the given facts.
-
-    :param facts: a dictionary of key, value pairs, where values are
-      Ansible result dictionaries.
-
-    :returns: a dictionary of key, value pairs, where the values are
-      strings. They will either be JSON-encoded data, or the empty
-      string on errors.
-    """
-    result = facts.copy()
-
-    # Note: we do NOT support transitive dependencies. If those are
-    # needed, this is the place to change.
-    for key, value in facts.items():
-        processor = PROCESSORS.get(key)
-        if not processor:
-            continue
-
-        # Use StopIteration to let the inner for loop continue the
-        # outer for loop.
-        try:
-            for dep in getattr(processor, 'DEPS', []):
-                if dep not in facts or \
-                   not facts[dep] or \
-                   isinstance(facts[dep], Exception):
-                    logger.error('Fact %s missing dependency %s',
-                                 key, dep)
-                    result[key] = NO_DATA
-                    raise StopIteration()
-        except StopIteration:
-            continue
-
-        if value.get('skipped', False):
-            logger.error('Fact %s skipped, no results', key)
-            result[key] = NO_DATA
-            continue
-
-        if value.get('rc', 0) and \
-           not getattr(processor, 'RETURN_CODE_ANY', False):
-            logger.error('Remote command returned %s', value['stdout'])
-            result[key] = NO_DATA
-            continue
-
-        try:
-            processor_out = processor.process(value)
-        except Exception as ex:  # pylint: disable=broad-except
-            logger.error('Processor returned %s', str(ex))
-            result[key] = NO_DATA
-            continue
-
-        # https://docs.python.org/3/library/json.html suggests that
-        # these separators will give the most compact representation.
-        result[key] = json.dumps(processor_out, separators=(',', ':'))
-
-    return result
-
-
-class ProcessorMeta(abc.ABCMeta):
-    """Metaclass to automatically register Processors."""
-
-    def __init__(cls, name, bases, dct):
-        """Register cls in the PROCESSORS dictionary."""
-        if 'KEY' not in dct:
-            raise Exception('Processor {0} does not have a KEY'.format(name))
-
-        # Setting a falsey KEY means "yes, I did this on purpose, I
-        # really don't want this class registered."
-        if dct['KEY']:
-            PROCESSORS[dct['KEY']] = cls
-        super().__init__(name, bases, dct)
-
-
 # pylint: disable=too-few-public-methods
-class Processor(object, metaclass=ProcessorMeta):
-    """A class to process the output of an Ansible task."""
-
-    # Special fields:
-    #   KEY: a string. Should match the Ansible fact name that this
-    #     will process (required).
-    #   DEPS: a list of keys of other facts that this one depends
-    #     on. If one of the deps had an error, we will show a "missing
-    #     dependency" message rather than an error for this fact
-    #     (optional).
-    #   RETURN_CODE_ANY: if True, process() will pass results with
-    #     non-zero return code to this processor. If False, process()
-    #     will emit a default error message if the shell command has
-    #     non-zero return code (optional, defaults to False).
-    #   process(output): a static method. Process the output of the
-    #     task (required).
-
-    KEY = None
-
-    @staticmethod
-    def process(output):
-        """Process Ansible output.
-
-        :param output: an Ansible output dictionary. This typically
-          has members 'rc', 'stdout', and 'stdout_lines', but may
-          instead have a member 'results', which is an array of dicts
-          that each have 'rc', 'stdout', 'stdout_lines', and
-          'item'.
-
-        :returns: a Python object that represents output. Returns
-          NO_DATA in case of error.
-        """
-        raise NotImplementedError()
-
 
 # #### Processors ####
 
@@ -143,7 +22,7 @@ FIND_WARNING = 'find: WARNING: Hard link count is wrong for /proc: this may' \
                ' be a bug in your filesystem driver.'
 
 
-class ProcessJbossEapRunningPaths(Processor):
+class ProcessJbossEapRunningPaths(process.Processor):
     """Process a list of JBoss EAP processes."""
 
     KEY = 'jboss_eap_running_paths'
@@ -155,12 +34,12 @@ class ProcessJbossEapRunningPaths(Processor):
         """Just preserve the output, except for a known issue."""
         if FIND_WARNING in output['stdout']:
             logging.error('Find command failed')
-            return NO_DATA
+            return process.NO_DATA
 
         return output['stdout'].strip()
 
 
-class ProcessFindJboss(Processor):
+class ProcessFindJboss(process.Processor):
     """Process the results of a find command."""
 
     KEY = 'jboss_eap_find_jboss_modules_jar'
@@ -171,7 +50,7 @@ class ProcessFindJboss(Processor):
         return output['stdout_lines']
 
 
-class ProcessIdUJboss(Processor):
+class ProcessIdUJboss(process.Processor):
     """Process the results of 'id -u jboss'."""
 
     KEY = 'jboss_eap_id_jboss'
@@ -189,10 +68,10 @@ class ProcessIdUJboss(Processor):
             return False
 
         logging.error('id: unexpected output %s', plain_output)
-        return NO_DATA
+        return process.NO_DATA
 
 
-class ProcessJbossEapCommonFiles(Processor):
+class ProcessJbossEapCommonFiles(process.Processor):
     """Process the output of 'test -e ...'."""
 
     KEY = 'jboss_eap_common_files'
@@ -215,7 +94,7 @@ class ProcessJbossEapCommonFiles(Processor):
         return out_list
 
 
-class ProcessJbossEapProcesses(Processor):
+class ProcessJbossEapProcesses(process.Processor):
     """Process the output of a process search."""
 
     KEY = 'jboss_eap_processes'
@@ -235,12 +114,12 @@ class ProcessJbossEapProcesses(Processor):
 
         if num_procs < 2:
             logging.error('Bad result from ps (%s processes)', num_procs)
-            return NO_DATA
+            return process.NO_DATA
 
         return num_procs - 2
 
 
-class ProcessJbossEapPackages(Processor):
+class ProcessJbossEapPackages(process.Processor):
     """Process the output of an rpm query."""
 
     KEY = 'jboss_eap_packages'
@@ -251,7 +130,7 @@ class ProcessJbossEapPackages(Processor):
         return len(output['stdout_lines'])
 
 
-class ProcessJbossEapLocate(Processor):
+class ProcessJbossEapLocate(process.Processor):
     """Process the output of 'locate jboss-modules.jar'."""
 
     KEY = 'jboss_eap_locate_jboss_modules_jar'
@@ -264,7 +143,7 @@ class ProcessJbossEapLocate(Processor):
         return output['stdout_lines']
 
 
-class InitLineFinder(Processor):
+class InitLineFinder(process.Processor):
     """Process the output of an init system.
 
     For both chkconfig and systemctl list-unit-files, we look for
@@ -302,7 +181,7 @@ class ProcessJbossEapSystemctl(InitLineFinder):
     KEY = 'jboss_eap_systemctl_unit_files'
 
 
-class IndicatorFileFinder(Processor):
+class IndicatorFileFinder(process.Processor):
     """Look for indicator files in the output of many 'ls -1's.
 
     Use by subclassing and defining a class variable INDICATOR_FILES,
@@ -338,7 +217,7 @@ class IndicatorFileFinder(Processor):
         return results
 
 
-class CatResultsProcessor(Processor):
+class CatResultsProcessor(process.Processor):
     """Look for 'Red Hat' in the output of many 'cat's.
 
     Use by making subclasses with their own KEYs.
